Create client and server libraries
Update protobuf version
Install third-party libraries
